Refactor Sidebar index.tsx - separate components into their own file
