+++ interface/app/$libraryId/Layout/Sidebar/IsRunningJob.tsx
@@ -1,0 +1,13 @@
+import { CheckCircle } from 'phosphor-react';
+import { Loader } from '@sd/ui';
+import { useLibraryQuery } from '~/../packages/client/src';
+
+export default () => {
+	const { data: isRunningJob } = useLibraryQuery(['jobs.isRunning']);
+
+	return isRunningJob ? (
+		<Loader className="h-[20px] w-[20px]" />
+	) : (
+		<CheckCircle className="w-5 h-5" />
+	);
+};
