+++ interface/app/$libraryId/Layout/Sidebar/Section.tsx
@@ -7,8 +7,8 @@ export default (
 		actionArea?: React.ReactNode;
 	}>
 ) => (
-	<div className="group mt-5">
-		<div className="mb-1 flex items-center justify-between">
+	<div className="mt-5 group">
+		<div className="flex items-center justify-between mb-1">
 			<CategoryHeading className="ml-1">{props.name}</CategoryHeading>
 			<div className="text-ink-faint opacity-0 transition-all duration-300 hover:!opacity-100 group-hover:opacity-30">
 				{props.actionArea}
