+++ interface/app/$libraryId/Layout/Sidebar/helpers.ts
@@ -1,0 +1,4 @@
+import { OperatingSystem } from "~/util/Platform";
+
+export const macOnly = (platform: OperatingSystem | undefined, classnames: string) =>
+	platform === 'macOS' ? classnames : '';
